# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Painting Mistery API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar roles",
                "operationId": "219aa99f0ea40b14fbb5f3b9dadc6ad1",
                "responses": {
                    "200": {
                        "description": "Lista de roles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear rol",
                "operationId": "1a4844f88745b12f75ecda89bd20fc13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Control total del sistema"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener rol en especifico",
                "operationId": "2d60f8015c5031ab0253f9fcab7e9adf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar rol",
                "operationId": "610553774a54e7dadc84dfeaa45d043e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Administrador"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Actualización de descripción"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "400": {
                        "description": "Sin datos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar rol",
                "operationId": "5bb912c54b58cb39916f46cb9b73200b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            }
        },
...
        "/tipo-iva": {
            "get": {
                "tags": [
                    "Tipo IVA"
                ],
                "summary": "Listar tipos de IVA",
                "operationId": "e946219dc3f84bc2238843e7a95a627d",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de IVA"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipo IVA"
                ],
                "summary": "Crear tipo de IVA",
                "operationId": "4432b3565f0f48e159db78d3ab4f6220",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descripcion",
                                    "porcentaje"
                                ],
                                "properties": {
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "porcentaje": {
                                        "type": "number",
                                        "example": 19
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/tipo-iva/{id}": {
            "get": {
                "tags": [
                    "Tipo IVA"
                ],
                "summary": "Obtener tipo de IVA",
                "operationId": "0bd9d8661ae29dbc33d85e144e8387b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipo IVA"
                ],
                "summary": "Actualizar tipo de IVA",
                "operationId": "7c8bc5f0392a8195dc30ed05552a2fdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "porcentaje": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipo IVA"
                ],
                "summary": "Eliminar tipo de IVA",
                "operationId": "7aa22f77cc50fa55d98290797b54d703",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
...
    },
...
}